Testing wrong java version fix.
Neither JAVA_HOME or JAVA_PATH (both the env bash var or the command-line option of braker.pl) were reaclling java version 1.8, even if it was installed. It was due to missing the braker.pl $JAVA_PATH variable on the $cmdString thagt was set to test for the java version. With this patchprobably you do not need to mess with update-alternatives, or other approaches to solve the issue of:
```
# Thu Sep 15 12:02:44 2022: Found environment variable $JAVA_PATH. Setting $JAVA_PATH to /usr/lib/jvm/java-8-openjdk-amd64/jre/bin
# Thu Sep 15 12:02:45 2022 ERROR: in file /usr/local/bin/braker.pl at line 3185
You have installed java version 17.0
. GUSHR requires version 1.8!
You can switch between java versions on your system with:
sudo update-alternatives --config java
```